First version of automatic interface generator
The goal is still far ahead, lots of work needs to be done, and testing. But,
so far, the script reads class.h, extract the name of the modules, read their
header file, create the classy.pyx with the proper fields, create wrapper
around the init function.

It essentially should create a low-level interface to CLASS, allowing for
writing a new version of the wrapper, in pure Python, that will not need to be
recompile every time we change it, should be much more readable and
user-friendly. Most importantly, this means that in the long term, any change
to the CLASS source code will be retransmitted to the wrapper seamlessly,
without human intervention (or so one can hope).