Renamed _struct_cleanup to struct_cleanup in the wrapper
Indeed, this function was being called by Monte Python. Also, the argument of
this function has been removed because it was always called with self.ncp. If
there was a need for it having another argument, a revert would be possible.

However, from a greater perspective, that this function does not need any
argument would make the wrapping of another cosmological code much easier from
the point of view of Monte Python.